+++ src/main/java/beans/entity/Address.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entities;
+package beans.entity;
 
 import java.io.Serializable;
 import java.util.Collection;
@@ -237,7 +237,7 @@ public class Address implements Serializable {
 
     @Override
     public String toString() {
-        return "entities.Address[ id=" + id + " ]";
+        return "beans.entity.Address[ id=" + id + " ]";
     }
     
 }
