+++ src/main/java/beans/entity/Category.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entities;
+package beans.entity;
 
 import java.io.Serializable;
 import java.util.Collection;
@@ -111,7 +111,7 @@ public class Category implements Serializable {
 
     @Override
     public String toString() {
-        return "entities.Category[ id=" + id + " ]";
+        return "beans.entity.Category[ id=" + id + " ]";
     }
     
 }
