+++ src/main/java/beans/entity/Company.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entities;
+package beans.entity;
 
 import java.io.Serializable;
 import java.util.Collection;
@@ -196,7 +196,7 @@ public class Company implements Serializable {
 
     @Override
     public String toString() {
-        return "entities.Company[ id=" + id + " ]";
+        return "beans.entity.Company[ id=" + id + " ]";
     }
     
 }
