+++ src/main/java/beans/entity/Invoice.java
@@ -3,11 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entities;
+package beans.entity;
 
 import java.io.Serializable;
 import java.util.Collection;
-import java.util.Date;
 import javax.persistence.Basic;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -131,7 +130,7 @@ public class Invoice implements Serializable {
 
     @Override
     public String toString() {
-        return "entities.Invoice[ id=" + id + " ]";
+        return "beans.entity.Invoice[ id=" + id + " ]";
     }
     
 }
