+++ src/main/java/beans/entity/PaymentMethod.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entities;
+package beans.entity;
 
 import java.io.Serializable;
 import javax.persistence.Basic;
@@ -110,7 +110,7 @@ public class PaymentMethod implements Serializable {
 
     @Override
     public String toString() {
-        return "entities.PaymentMethod[ id=" + id + " ]";
+        return "beans.entity.PaymentMethod[ id=" + id + " ]";
     }
     
 }
