+++ src/main/java/beans/session/CategoryFacade.java
@@ -3,9 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sessionBeans;
+package beans.session;
 
-import entityBeans.Company;
+import entityBeans.Category;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
@@ -14,7 +14,7 @@ import javax.persistence.PersistenceContext;
  * @author douwejongeneel
  */
 @javax.ejb.Stateless
-public class CompanyFacade extends AbstractFacade<Company> {
+public class CategoryFacade extends AbstractFacade<Category> {
 
     @PersistenceContext(unitName = "rsvier_workshop3.rest_war_1.0-SNAPSHOTPU")
     private EntityManager em;
@@ -24,8 +24,8 @@ public class CompanyFacade extends AbstractFacade<Company> {
         return em;
     }
 
-    public CompanyFacade() {
-        super(Company.class);
+    public CategoryFacade() {
+        super(Category.class);
     }
     
 }
