+++ src/main/java/beans/session/PaymentFacade.java
@@ -3,9 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sessionBeans;
+package beans.session;
 
-import entityBeans.Invoice;
+import entityBeans.Payment;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
@@ -14,7 +14,7 @@ import javax.persistence.PersistenceContext;
  * @author douwejongeneel
  */
 @javax.ejb.Stateless
-public class InvoiceFacade extends AbstractFacade<Invoice> {
+public class PaymentFacade extends AbstractFacade<Payment> {
 
     @PersistenceContext(unitName = "rsvier_workshop3.rest_war_1.0-SNAPSHOTPU")
     private EntityManager em;
@@ -24,8 +24,8 @@ public class InvoiceFacade extends AbstractFacade<Invoice> {
         return em;
     }
 
-    public InvoiceFacade() {
-        super(Invoice.class);
+    public PaymentFacade() {
+        super(Payment.class);
     }
     
 }
